Add CUDA + PyTorch (#31)
* Add CUDA + PyTorch

Assumes CUDA 12.0 is installed. Might work with other versions.

Note that installing pytorch is by far the longest
part of this process. By using an image with pytorch pre-installed,
this could build much faster. However, copying torch from an image
would be pretty hard, so it might be worth copying/installing ROS
on top of a pre-built torch container.

* Build with/without GPU via arguments

Allows for disabling (--no-gpu) GPU, or forcing
(--force-gpu) GPU on platforms without a GPU.

Development environment defaults to use no GPU if
nvidia container toolkit is not present. Production
will throw a warning and fail if this is the case
and --no-gpu is not explicitly specified.

* Lock down all CUDA requirement versions

+++ setup_prod.sh
@@ -1,11 +1,71 @@
 #!/bin/bash
-# Run to spin up docker containers and set aliases
-docker kill $(docker ps -q) # kill all running containers
-docker compose build
-docker compose --env-file .env.prod up -d
+
+no_gpu=false
+force_gpu=false
+
+usage() {
+  echo "Usage: $0 [options]"
+  echo "Options:"
+  echo " --no-gpu        Disable CUDA support. Required to run on systems without Nvidia Container Toolkit."
+  echo " --force-gpu     Force a GPU build even if Nvidia Container Toolkit is not detected"
+}
+
+while [ $# -gt 0 ]; do
+  case $1 in
+    -h | --help)
+      usage
+      exit 0
+      ;;
+    --no-gpu)
+      no_gpu=true
+      ;;
+    --force-gpu)
+      force_gpu=true
+      ;;
+    *)
+      echo "Invalid option: $1" >&2
+      usage
+      exit 1
+      ;;
+  esac
+  shift
+done
+
+if $no_gpu && $force_gpu
+then
+  echo -e "\033[0;31mOptions --no-gpu and --force-gpu conflict. Use at most one of them.\033[0m"
+  exit 1
+fi
+
+if ! $no_gpu && ! $force_gpu && ! command -v nvidia-ctk &> /dev/null
+then
+  # In production, we should require it to be specified when we want no GPU.
+  echo -e "\033[0;31mNvidia Container Toolkit was not found.\033[0m"
+  echo -e "\033[0;31mRun with --no-gpu to build without GPU, or --force-gpu to force a GPU build.\033[0m"
+  exit 1
+fi
+
+#####################
+# Actual logic here #
+#####################
+
+if $no_gpu
+then
+  dockerfile="docker-compose-no-gpu.yml"
+else
+  dockerfile="docker-compose-gpu.yml"
+fi
+
+echo "Killing all containers..."
+docker kill $(docker ps -q)
+
+echo "Building containers..."
+docker compose -f $dockerfile build
+
+echo "Starting containers..."
+docker compose -f $dockerfile --env-file .env.prod up -d
 
 sleep 0.5
 
 echo "DEBUG: Buggy Docker Container Up!"
 echo "Run docker_exec in order to go into the Docker container"
-
